feat: refactor account resources
