+++ src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.container-image.group=quarkus-mp
+quarkus.container-image.name=account-service
+quarkus.kubernetes.name=account-service
